seqlevelsInUse {Seqinfo}	R Documentation

Get the sequence levels in use

Description

Get the sequence levels that are "used" by an object, that is, the sequence levels on which the object defines genomic ranges or features.

Usage

seqlevelsInUse(x)

Arguments

x
Any object containing sequence information i.e. with a seqinfo() component.

Value

The sequence levels in use in a character vector.

Author(s)

H. Pagès

See Also

The seqlevels getter and setter.

Seqinfo objects.

GRanges and GRangesList objects in the GenomicRanges package.

Examples

library(GenomicRanges)

gr <- GRanges(rep(c("chr2", "chr3", "chrM"), 2), IRanges(1:6, 10))

## Add new seqlevels:
seqlevels(gr) <- c("chr1", seqlevels(gr), "chr4")
seqlevels(gr)
seqlevelsInUse(gr)

## Drop all unused seqlevels:
seqlevels(gr) <- seqlevelsInUse(gr)

[Package Seqinfo version 1.0.0 Index]