sortSeqlevels {Seqinfo}	R Documentation

Sort the sequence levels of an object

Description

A generic function and methods to sort the sequence levels of an object in "natural" order.

Usage

sortSeqlevels(x, X.is.sexchrom=NA)

Arguments

x
Any object containing sequence information i.e. with a seqinfo() component.

X.is.sexchrom
A logical indicating whether X refers to the sexual chromosome or to chromosome with Roman Numeral X. If NA, sortSeqlevels does its best to "guess".

Value

The input object x with its seqlevels sorted.

Author(s)

H. Pagès

See Also

The seqlevels getter and setter.

rankSeqlevels, on which sortSeqlevels is based.

Seqinfo objects.

GRanges and GRangesList objects in the GenomicRanges package.

Examples

sortSeqlevels(c("11", "Y", "1", "10", "9", "M", "2"))

seqlevels <- c("chrXI", "chrY", "chrI", "chrX", "chrIX", "chrM", "chrII")
sortSeqlevels(seqlevels)
sortSeqlevels(seqlevels, X.is.sexchrom=TRUE)
sortSeqlevels(seqlevels, X.is.sexchrom=FALSE)

seqlevels <- c("chr2RHet", "chr4", "chrUextra", "chrYHet",
               "chrM", "chrXHet", "chr2LHet", "chrU",
               "chr3L", "chr3R", "chr2R", "chrX")
sortSeqlevels(seqlevels)

library(GenomicRanges)
gr <- GRanges()
seqlevels(gr) <- seqlevels
sortSeqlevels(gr)

[Package Seqinfo version 1.0.0 Index]